feat: move to modules and more
